LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scriptnonce-allowed-expected.txt

ALERT: PASS (1/1)
This tests the effect of a valid script-nonce value. It passes if no console warning is visible, and the alert() is executed.
